+++ services/madoc-ts/src/extensions/capture-models/CollectionExplorer/CollectionExplorer.editor.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const CollectionExplorerEditor = () => {
+  return null;
+};
+
+export default CollectionExplorerEditor;